Create a JS keyboard panel
This is done so that we can use https://github.com/vadimdemedes/ink
to render the CLI output, specifically so that we can overwrite the existing
output rather than continue to print additional lines, making it hard to read
the output and pushing the user’s input ever downward.

I tried really hard to improve the existing Python KeyboardPanel by using
curses and urwid, without luck.