Fix the timeout bar not showing up at 100%
It showed up with a delay previously because we never emitted a change event
for the initial value.

The `start` call is moved to `assignControl` because we only observe the
change events on commands that are part of `this._commands`, and rather
than change the ordering of the calls to assign the control to the panel and
add to `this._commands`, it’s more explicit to just call `start` after both.