Time out commands before the game starts
So that if a player can't figure out how to perform a command, they're not
prevented from joining the game, they can just wait for the next command. To
give them time to figure out how the game works, they get double the time to
perform before the game starts.